test(tldraw): consolidate held-key shortcut tests into keyboardShortcuts.test.tsx
